requirements.txt
streamlit==1.31.1
tensorflow==2.12.0
opencv-python-headless==4.8.1.78
numpy==1.23.5
pillow==10.2.0
gdown==5.1.0

